fix(parsing): no longer JSON clone the returned data
Fixes https://github.com/Swydo/blaze-apollo/issues/8.

We used to clone the data because developers could modify the returned
data object. This would prevent the equals function to work. Now that
the Apollo client returns frozen (immutable) data, we no longer need to
clone it, because it’s impossible to modify the original data. Which is
great.